Create and store dictionaries

NOTE: DO NOT USE THIS FUNCTION DIRECTLY. INSTEAD, USE THE "AT-SCALE" BASH SCRIPT, WHICH IS DESCRIBED HERE: https://timothy-barry.github.io/sceptre/articles/sceptre-at-scale.html

create_and_store_dictionaries(
  gene_gRNA_pairs,
  gene_precomp_dir,
  gRNA_precomp_dir,
  results_dir,
  pod_sizes
)

Arguments

gene_gRNA_pairs
a data frame containing a the gene-gRNA pairs to analyze using sceptre. The column names should include "gene_id" and "gRNA_id."

gene_precomp_dir
the directory in which to store the gene precomputations

gRNA_precomp_dir
the directory in which to store the gRNA precomputations

results_dir
the directory in which to store the results

pod_sizes
an integer vector with three named elements: gRNA, gene, and pair. These elements give the sizes of the respective "pods."

Value

an integer vector containing the number of pods in the gene, gRNA, and pairs dictionaries.

Details

We require a few bookkeeping files (that we call "dictionaries") for the gRNA and gene precomputation steps. This file creates those dictionaries and stores them in the appropriate locations on disk.