Generate all pairs

The generate_all_pairs() function generates all possible response-gRNA group pairs that can be constructed from a given response_matrix and grna_group_data_frame.

generate_all_pairs(response_matrix, grna_group_data_frame)

Arguments

response_matrix

(required) a matrix of raw expression counts. The responses (e.g., genes or proteins) should be in the rows, and the cells should be in the columns. The row names should be the unique IDs of the responses. The matrix can be a standard (dense) matrix or a sparse matrix of class dgCMatrix, dgRMatrix, or dgTMatrix.

grna_group_data_frame

(required) a data frame specifying the group to which each individual gRNA belongs. This data frame should contain columns grna_id and grna_group, with grna_group specifying the group membership of a given gRNA in the grna_id column. Typically, gRNAs that target the same site are grouped into the same gRNA group. IMPORTANT: Non-targeting gRNAs should be assigned a gRNA group label of "non-targeting".

Value

a data frame containing the columns response_id and grna_group in which each response is mapped to the entire set of targeting gRNA groups.

Examples

data(response_matrix_lowmoi)
data(grna_group_data_frame_lowmoi)
response_grna_group_pairs <- generate_all_pairs(response_matrix_lowmoi,
grna_group_data_frame_lowmoi)