Plot calibration result

The plot_calibration_result() function helps to visualize the results of a calibration check.

plot_calibration_result(
  calibration_result,
  alpha = 0.1,
  multiple_testing_correction = "BH",
  return_indiv_plots = FALSE
)

Arguments

calibration_result

(required) the output of a call to run_sceptre_lowmoi with calibration_check set to TRUE

alpha

(optional; default 0.1) the nominal type-I error control level

multiple_testing_correction

(optional; default "BH") the multiple testing correction method (one of "BH", "bonferroni", "BY", or "holm")

return_indiv_plots

(optional; default FALSE) return a single combined plot (TRUE) or the individual plots in a list (FALSE)?

Value

a single cowplot object containing the combined panels (if return_indiv_plots is set to TRUE) or a list of the individual panels (if return_indiv_plots is set to FALSE).

Details

The plot contains four panels.

upper left: a Q-Q plot of the p-values on an untransformed scale. The p-values should lie along the diagonal line.

upper right: a Q-Q plot of the p-values on a negative log-10 transformed scale. The p-values should lie along the diagonal line, and the majority of the p-values should fall within the gray confidence band.

bottom left: a histogram of the estimated log-2 fold changes in expression. The histogram should be roughly symmetric and centered at zero.

bottom right: a text box displaying (i) the number of false discoveries made on the negative control pairs and (ii) the mean estimated log-fold change. The number of false discoveries ideally should be zero, although a small, positive integer (such as 1, 2, or 3) also is OK. The mean estimated log-fold change should be a numeric value close to zero.

Note

The untransformed and transformed Q-Q plots are both informative: the former enables us to see the bulk of the p-value distribution, while the latter enables visualize the tail.

The number of false discoveries depends both on alpha and multiple_testing_correction. The default value of these arguments is 0.1 and "BH". This corresponds to a BH correction at nominal false discovery rate (FDR) 0.1.

Technical point: when applying BH at level 0.1 to a collection of strictly null p-values, BH controls family-wise error rate (FWER) at level 0.1 as well as FDR at level 0.1. FWER is the probability of making one or more false discoveries. Thus, with probability 0.9, the number of rejections that BH makes on (well-calibrated) null p-values at level 0.1 is 0. This implies that sceptre (or any method for that matter) should make about zero false discoveries on negative control p-values data after applying a BH correction at level 0.1.

Examples

# See the example in the run_sceptre_lowmoi help file.
?run_sceptre_lowmoi