Plot covariates

plot_covariates() creates histograms of several cell-specific covariates: response_n_nonzero, response_n_umis, and (if applicable) response_p_mito. To help guide the selection of QC thresholds, plot_covariates() plots candidate QC thresholds as vertical lines on the histograms.

Usage

plot_covariates(
  sceptre_object,
  response_n_umis_range = c(0.01, 0.99),
  response_n_nonzero_range = c(0.01, 0.99),
  p_mito_threshold = 0.2,
  return_indiv_plots = FALSE
)

Arguments

sceptre_object

any initialized sceptre_object. plot_covariates() can be called at any point in the pipeline after import_data().

response_n_umis_range

(optional; default c(0.01, 0.99)) a length-2 vector of quantiles indicating the location at which to draw vertical lines on the response_n_umis histogram

response_n_nonzero_range

(optional; default c(0.01, 0.99)) a length-2 vector of quantiles indicating the location at which to draw vertical lines on the response_n_nonzero histogram

p_mito_threshold

(optional; default 0.2) a single numeric value in the interval [0,1] specifying the location at which to draw a vertical line on the response_p_mito histogram. Note that p_mito_threshold is an absolute number rather than a percentile.

return_indiv_plots

(optional; default FALSE) if FALSE, then a list of ggplot objects is returned; if TRUE then a single cowplot object is returned.

Value

a single cowplot object containing the combined panels (if return_indiv_plots is set to TRUE) or a list of the individual panels (if return_indiv_plots is set to FALSE)

Examples

# A full example can be found at ?sceptre