Plot result

For a given gRNA-gene pair analyzed by sceptre, plots the resampled test statistics alongside the "ground truth" test statistic derived from the raw data.

plot_result(row)

Arguments

row

single row of the data frame outputted by run_sceptre_high_moi, when full_output is set to TRUE

Value

a ggplot2 object containing the plot

Examples

# RUN THE METHOD
set.seed(1)
data(gene_matrix)
data(gRNA_matrix)
data(covariate_matrix)
data(gRNA_groups_table)
data(gene_gRNA_group_pairs)
combined_perturbation_matrix <- threshold_gRNA_matrix(gRNA_matrix) %>%
combine_perturbations(gRNA_groups_table)
#> Error in threshold_gRNA_matrix(gRNA_matrix) %>% combine_perturbations(gRNA_groups_table): could not find function "%>%"
gene_gRNA_group_pairs <- gene_gRNA_group_pairs %>% sample_n(1)
#> Error in gene_gRNA_group_pairs %>% sample_n(1): could not find function "%>%"
result <- run_sceptre_high_moi(gene_matrix = gene_matrix,
combined_perturbation_matrix = combined_perturbation_matrix,
covariate_matrix = covariate_matrix,
gene_gRNA_group_pairs = gene_gRNA_group_pairs,
side = "left", parallel = FALSE, full_output = TRUE)
#> Check /var/folders/h6/pt59hmwd7rqb3jqfzwh0hlth0000gn/T//Rtmpjdo5Sf/logs for more detailed status updates.
#> Running checks and setting up directory structure. 
#> Error in run_sceptre_high_moi(gene_matrix = gene_matrix, combined_perturbation_matrix = combined_perturbation_matrix,     covariate_matrix = covariate_matrix, gene_gRNA_group_pairs = gene_gRNA_group_pairs,     side = "left", parallel = FALSE, full_output = TRUE): object 'combined_perturbation_matrix' not found

# CREATE THE PLOT
plot_result(result[1,])
#> Error in dplyr::select(., dplyr::starts_with("z_null_")): object 'result' not found