Plot run QC

plot_run_qc() creates a visualization of the outcome of the QC step. See Section 4 of the introductory chapter in the manual for guidance on interpreting this visual.

Usage

plot_run_qc(
  sceptre_object,
  downsample_pairs = 10000L,
  point_size = 0.55,
  transparency = 0.8,
  return_indiv_plots = FALSE
)

Arguments

sceptre_object

a sceptre_object that has had run_qc() called on it

downsample_pairs

(optional; default 10000) the maximum number of points to plot in the lower panel of the figure (i.e., the "pairwise QC" plot)

point_size

(optional; default 0.55) the size of the individual points in the plot

transparency

(optional; default 0.8) the transparency of the individual points in the plot

return_indiv_plots

(optional; default FALSE) if FALSE then a list of ggplot is returned; if TRUE then a single cowplot object is returned.

Value

a single cowplot object containing the combined panels (if return_indiv_plots is set to TRUE) or a list of the individual panels (if return_indiv_plots is set to FALSE)

Examples

# A full example can be found at ?sceptre;
# `plot_run_qc()` is dispatched when
# `plot()` is called on the `sceptre_object`
# in step 4 (the run qc step).