Run gRNA-gene pair analysis at scale

Runs the gene-gRRNA pair ananysis across an entire pod of gene-gRNA pairs.

run_gRNA_gene_pair_analysis_at_scale(
  pod_id,
  gene_precomp_dir,
  gRNA_precomp_dir,
  results_dir,
  log_dir,
  gene_matrix,
  gRNA_matrix,
  covariate_matrix,
  regularization_amount,
  side,
  B,
  full_output
)

Arguments

pod_id

id of the pod to compute

gene_precomp_dir

the directory containing the results of the gene precomputation

gRNA_precomp_dir

the directory containing the results of the gRNA precomputation

results_dir

directory in which to store the results

log_dir

(optional) directory in which to sink the log file

gene_matrix

a gene-by cell expression matrix; the rows (i.e., gene IDs) and columns (i.e., cell barcodes) should be named

gRNA_matrix

a gRNA-by cell expression matrix; the rows (i.e., gRNA IDs) and columns (i.e., cell barcodes) should be named

covariate_matrix

the cell-specific matrix of technical factors, ideally containing the following covariates: log-transformed gene library size (numeric), log-transformed gRNA library size (numeric), percent mitochondrial reads (numeric), and batch (factor). The rows (i.e., cell barcodes) should be named

regularization_amount

non-negative number specifying the amount of regularization to apply to the negative binomial dispersion parameter estimates

side

sidedness of the test; one of "both," "left," and "right"

B

number of resamples to draw for the conditional randomization test

full_output

return the full output (TRUE) or a simplified, reduced output (FALSE)?