Run SCEPTRE on a gRNA-gene pair

Runs SCEPTRE on a given gRNA-gene pair. The function requires as arguments the gene expression vector, the gRNA indicator vector, and the covariate matrix.

run_sceptre_gRNA_gene_pair(
  expressions,
  gRNA_indicators,
  covariate_matrix,
  side = "left",
  gene_precomp_size = NULL,
  B = 500,
  seed = NULL,
  reduced_output = TRUE,
  verbose = TRUE
)

Arguments

expressions
a vector a gene expressions

gRNA_indicators
a vector of gRNA indicators

covariate_matrix
the data frame of cell-specific covariates

side
(optional; default "left") side of the test; one of "left," "right," and "both."

gene_precomp_size
(optional) the pre-computed size of the gene NB distribution; if not supplied, will be estimated

B
(optional; default 500) number of conditional randomization test resamples

seed
(optional; default 4) seed of the random number generator

reduced_output
(optional; default TRUE) return the reduced output?

verbose
(optional; default TRUE) print status updates?

Value

If reduced_output is TRUE, a data frame containing the following columns: p-value, skew-t fit success, skew-t fit MLEs (xi, omega, alpha, nu), the original signed test statistic, and the number of successful resamples; if reduced_output is FALSE, a list containing all of the above, plus the vector of resampled test statistics.

Examples

data(expressions)
data(gRNA_indicators)
data(covariate_matrix)
result <- run_sceptre_gRNA_gene_pair(expressions,
gRNA_indicators,
covariate_matrix,
verbose = FALSE)
result

#>     p_value skew_t_fit_success        xi    omega      alpha       nu
#> 1 0.1952004               TRUE 0.8196897 1.021334 -0.3667235 59745.98
#>      z_value n_successful_resamples
#> 1 -0.3037842                    500