Run sceptre on low MOI data

The function run_sceptre_lowmoi() runs sceptre on low multiplicity of infection (MOI) single-cell CRISPR screen data. sceptre tests for association (and estimates the log fold change in expression) for a user-specified set of response-gRNA group pairs. Applying sceptre consists of three main steps: (1) setting up the analysis, (2) carrying out a "calibration check" to verify that sceptre controls the rate of false discoveries on negative control data, and (3) conducting the "discovery analysis" to discover regulatory relationships.

Usage

run_sceptre_lowmoi(
  response_matrix,
  grna_matrix,
  covariate_data_frame,
  grna_group_data_frame,
  formula_object,
  response_grna_group_pairs,
  calibration_check,
  n_nonzero_trt_thresh = 7L,
  n_nonzero_cntrl_thresh = 7L,
  discovery_test_stat = "exact",
  return_debugging_metrics = FALSE,
  return_resampling_dist = FALSE,
  fit_skew_normal = TRUE,
  calibration_group_size = NULL,
  n_calibration_pairs = NULL,
  B1 = 499L,
  B2 = 4999L,
  B3 = 24999L,
  regression_method = "poisson_glm",
  print_progress = TRUE
)

Arguments

response_matrix

(required) a matrix of raw expression counts. The responses (e.g., genes or proteins) should be in the rows, and the cells should be in the columns. The row names should be the unique IDs of the responses. The matrix can be a standard (dense) matrix or a sparse matrix of class dgCMatrix, dgRMatrix, or dgTMatrix.

grna_matrix

(required) a matrix of gRNA expression counts. The gRNAs should be in the rows, and the cells should be in the columns. The row names should be the unique IDs of the gRNAs. The matrix can be a standard (dense) matrix or a sparse matrix of class dgCMatrix, dgRMatrix, or dgTMatrix. (See note about passing logical gRNA matrices.)

covariate_data_frame

(required) a data frame containing the cell-specific covariates (e.g., sequencing batch, total UMI count, etc.), with cells in the rows and covariates in the columns. The column names should be the names of the covariates.

grna_group_data_frame

(required) a data frame specifying the group to which each individual gRNA belongs. This data frame should contain columns grna_id and grna_group, with grna_group specifying the group membership of a given gRNA in the grna_id column. Typically, gRNAs that target the same site are grouped into the same gRNA group. IMPORTANT: Non-targeting gRNAs should be assigned a gRNA group label of "non-targeting".

formula_object

(required) an R formula object specifying how sceptre is to adjust for the covariates (in the covariate_data_frame).

response_grna_group_pairs

(required) a data frame specifying the set of response-gRNA group pairs to test for association. The data frame should contain columns response_id and grna_group.

calibration_check

(required) a logical (i.e., TRUE/FALSE) value indicating whether to run a calibration check (TRUE) or a discovery analysis (FALSE). See the "Details" section for a more complete description of the calibration and discovery analyses.

n_nonzero_trt_thresh

(optional; moderate importance; default 7) For a given response-gRNA group pair, n_nonzero_trt is the number of "treatment cells" (i.e., cells that have received a targeting gRNA) with nonzero expression. sceptre filters for response-gRNA group pairs with an n_nonzero_trt value equal to or greater than n_nonzero_trt_thresh.

n_nonzero_cntrl_thresh

(optional; moderate importance; default 7) n_nonzero_cntrl is the number of "control cells" (i.e., cells that have received an NT gRNA) with nonzero expression. sceptre filters for pairs with a n_nonzero_control value equal to or greater than n_nonzero_control_thresh.

discovery_test_stat

(optional; lower importance; default "exact") the test statistic to use in the discovery analysis, either "exact" or "approximate".

return_debugging_metrics

(optional; lower importance; default FALSE) return metrics that facilitate debugging calibration issues? (See "Note" for details.)

return_resampling_dist

(optional; lower importance; default FALSE) return the resampling distribution of the null test statistics?

fit_skew_normal

(optional; lower importance; default TRUE) fit a skew-normal distribution to the null distribution of test statistics and use this fitted distribution to compute a more accurate p-value?

calibration_group_size

(optional; lower importance; default NULL) the number of NT gRNAs to assign to each negative control gRNA group in the calibration check analysis. By default, sceptre sets calibration_group_size to the median group size of the targeting gRNAs.

n_calibration_pairs

(optional; lower importance; default NULL) the number of negative control pairs to analyze in the calibration check. By default, the number of negative control pairs that sceptre analyzes is equal to the number of discovery pairs (i.e., pairs specified in the data frame response_grna_group_pairs) that passes pairwise QC.

B1

(optional; lower importance; default 499) the number of null test statistics to compute in the first round of the permutation test.

B2

(optional; lower importance; default 4999) the number of null test statistics to compute the second round of the permutation test.

B3

(optional; lower importance; default 24999) the number of null test statistics to compute the third round of the permutation test.

regression_method

(optional; lower importance; default "poisson_glm") The regression method to use to regress out the covariates, either "poisson_glm" for a Poisson GLM or "nb_glm" for a negative binomial GLM.

print_progress

(optional; default TRUE) print the progress of the function?

Value

a data frame containing the following columns: response_id, grna_group, n_nonzero_trt, n_nonzero_cntrl, and p_value.

response_id: the response ID

grna_group: the gRNA group. When running a calibration check, grna_group is a negative control gRNA group constructed by grouping together individual NT gRNAs. When running a discovery analysis, by contrast, grna_group is a targeting gRNA group.

n_nonzero_trt: the number of treatment cells (i.e., cells that have received a targeting gRNA) with nonzero expression

n_nonzero_cntrl: the number of control cells (i.e., cells that have received a non-targeting gRNA) with nonzero expression

p_value: the sceptre p-value

log_2_fold_change: the sceptre-estimated log (base 2) fold change in expression

Rows are ordered according to p_value. Response-gRNA group pairs that fail to pass pairwise QC are assigned a p-value and log fold change estimate of NA.

If return_resampling_dist is set to TRUE, the null permutation distribution is returned, stored in columns named "z_null_1", ..., "z_null_B1". If return_debugging_metrics is set to TRUE, the additional columns sn_fit_used, round, and regression_method are included in the ouput data frame. These column are described in the "Notes" section.

Details

A typical sceptre analysis consists of three main steps. First, the user sets up the analysis, which entails specifying (i) the response expression matrix, (ii) the gRNA expression matrix, (iii) the cell covariate data frame, (iv) the gRNA group data frame, (v) the formula object, and (vi) the set of response-gRNA group pairs to test for association. Second, the user carries out the calibration check by passing the above objects to the run_sceptre_lowmoi function, setting calibration_check to TRUE. In carrying out the calibration check, sceptre automatically constructs a set of negative control response-gRNA group pairs to test for association. The negative control pairs are "matched" to the discovery pairs in three ways: (i) the negative control pairs are subjected to the same pair-wise QC as the discovery pairs, (ii) the number of negative control pairs equals the number of discovery pairs (after applying pair-wise QC), and (iii) the negative control pairs contain the same number of gRNAs per gRNA group as the discovery pairs. After running run_sceptre_lowmoi, the user verifies calibration of the p-values by plotting the calibration results via plot_calibration_result. Finally, the user runs a discovery analysis by again calling run_sceptre_lowmoi, this time setting calibration_check to FALSE. The user visualizes the discovery results via calls to compare_calibration_and_discovery_results and make_volcano_plot and obtains the final set of discoveries via a call to obtain_discovery_set.

Note

The notes are arranged roughly in order of most important to least important.

The function compute_cell_covariates can be used to help compute the covariate_data_frame, and the function generate_all_pairs can be used to help compute response_grna_group_pairs.

When constructing the formula_object, it is best practice to log-transform the integer count-based variables (e.g., total UMI count, number of expressed responses).

By default, sceptre assigns to each cell the gRNA with highest UMI count. Users may apply a different gRNA-to-cell assignment themselves, and then pass a logical grna_matrix (i.e., TRUE/FALSE) to sceptre instead of a count matrix. Each column should contain a single TRUE value, which indicates which gRNA has infected the given cell. The logical matrix should be a standard (dense) R matrix or a sparse matrix of type lgCMatrix, lgRMatrix, lgTMatrix.

sceptre uses three rounds of permutation testing to test a given pair. First, sceptre computes B1 null statistics and calculates an initial empirical p-value p1 using these null statistics. If p1 is promising (i.e., p1 < 0.02), then sceptre proceeds to round 2; otherwise, sceptre returns p1. In round 2, sceptre computes B2 null statistics and fits a skew normal distribution to these null statistics. If the skew-normal fit is good, then sceptre returns p2. Otherwise, sceptre proceeds to round 3. In round 3, sceptre computes B3 null statistics and calculates an empirical p-value p3 using these null statistics. sceptre then returns p3.

When return_debugging_metrics is set to TRUE, the columns sn_fit_used, round, and regression_method are included in the output data frame. sn_fit_used indicates whether a p-value was calculated via the skew-normal distribution; sn_fit_used is FALSE if (i) sceptre computes the p-value in round 1 or round 3 or if (ii) fit_skew_normal is set to FALSE by the user. round indicates the round (among rounds 1, 2, and 3) in which the p-value was computed. Finally, regression_method is a string indicating the method that was used to regress the gene expression vector onto the cell covariate matrix and estimate the negative binomial theta parameter.

When return_resampling_dist is set to TRUE, sceptre uses only a single round of permutation testing, returning B1 resampled statistics for each pair. return_resampling_dist should be set to TRUE for debugging purposes only.

The discovery_test_stat parameter controls the test statistic used. The "exact" statistic is more robust, while the "approximate" statistic is faster. In most cases the two test statistics yield similar results.

Examples

if (FALSE) {
library(Matrix)

# 0. load the data associated with the experiment
data(response_matrix_lowmoi) # response-by-cell expression matrix
data(grna_matrix_lowmoi) # gRNA-by-cell expression matrix
data(covariate_data_frame_lowmoi) # cell-by-covariate data frame
data(grna_group_data_frame_lowmoi) # gRNA group information

# 1. obtain the set of pairs to analyze
response_grna_group_pairs <- generate_all_pairs(response_matrix_lowmoi,
grna_group_data_frame_lowmoi)

# 2. set the formula object
formula_object <- formula(~log(response_n_umis) + log(response_n_nonzero) +
bio_rep + p_mito)

# 3. run the calibration check analysis (NOTE: calibration_check set to TRUE)
calibration_result <- run_sceptre_lowmoi(response_matrix = response_matrix_lowmoi,
grna_matrix = grna_matrix_lowmoi,
covariate_data_frame = covariate_data_frame_lowmoi,
grna_group_data_frame = grna_group_data_frame_lowmoi,
formula_object = formula_object,
response_grna_group_pairs = response_grna_group_pairs,
calibration_check = TRUE)

# 4. plot the calibration result to ensure adequate calibration
plot_calibration_result(calibration_result)

# 5. run the discovery analysis (NOTE: calibration_check set to FALSE)
discovery_result <- run_sceptre_lowmoi(response_matrix = response_matrix_lowmoi,
grna_matrix = grna_matrix_lowmoi,
covariate_data_frame = covariate_data_frame_lowmoi,
grna_group_data_frame = grna_group_data_frame_lowmoi,
formula_object = formula_object,
response_grna_group_pairs = response_grna_group_pairs,
calibration_check = FALSE)

# 6. compare discovery p-values to the negative control p-values; make a volcano plot
compare_calibration_and_discovery_results(calibration_result, discovery_result)
make_volcano_plot(discovery_result)

# 7. obtain the discovery set for downstream analysis
discovery_set <- obtain_discovery_set(discovery_result)
}